LayoutTests/fast/preloader/frame-src.html

<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpResourceResponseMIMETypes();
</script>
<script src="../../resources/js-test-pre.js"></script>
<script src=resources/non-existant.js></script>
<script>
    shouldBeTrue('internals.isPreloaded("resources/testFrame.html");');
</script>
</head>
<body>
<p>Tests that iframes are preloaded.</p>
<iframe src="resources/testFrame.html"></iframe>
</body>
</html>
